Creacion de la entidad de Productos
