+++ PupShops/back/pupshops/src/modules/products/products.repository.ts
@@ -1,47 +1,21 @@
-import { Column, Entity, PrimaryGeneratedColumn } from "typeorm";
+import { Injectable } from "@nestjs/common";
+import { InjectRepository } from "@nestjs/typeorm";
+import { Products } from "./entities/product.entity";
+import { Repository } from "typeorm";
 
 
 
+@Injectable()
+export class ProductsRepository{
+    constructor(@InjectRepository(Products) private productsRepository: Repository<Products>){}
 
-@Entity({
-    name: 'products',
-})
-export class Products{
-    @PrimaryGeneratedColumn('uuid')
-    id:string;
+    getProducts(){}
 
+    getOneProduct(){}
 
-    @Column({
-        type: 'varchar',
-        length: 100,
-        unique:true,
-        nullable:false,
-    })
-    name: string;
+    addProducts(){}
 
-    @Column({
-        type: 'text',
-        nullable: false,
-    })
-    description:string;
+    updateProducts(){}
 
-    @Column({
-        type: 'decimal',
-        precision: 10,
-        scale:2,
-        nullable:false,
-    })
-    price:number;
-
-    @Column({
-        type: 'int',
-        nullable:false
-    })
-    stock:number;
-
-    @Column({
-        type: 'text',
-        default: 'http'
-    })
-    imgUrl: string;
+    deleteProducts(){}
 }
