+++ PupShops/back/pupshops/src/modules/products/products.service.ts
@@ -1,26 +1,25 @@
 import { Injectable } from '@nestjs/common';
-import { CreateProductDto } from './dto/create-product.dto';
-import { UpdateProductDto } from './dto/update-product.dto';
+
 
 @Injectable()
 export class ProductsService {
-  create(createProductDto: CreateProductDto) {
+  createProduct() {
     return 'This action adds a new product';
   }
 
-  findAll() {
+  findAllProducts() {
     return `This action returns all products`;
   }
 
-  findOne(id: number) {
+  findOneProduct(id: number) {
     return `This action returns a #${id} product`;
   }
 
-  update(id: number, updateProductDto: UpdateProductDto) {
+  updateProduct(id: number) {
     return `This action updates a #${id} product`;
   }
 
-  remove(id: number) {
+  removeProduct(id: number) {
     return `This action removes a #${id} product`;
   }
 }
